hsi-3851 [BE] Corrección de tipo en la columna rol en la tabla CareLineRole

+++ back-end/hospital-api/src/main/java/net/pladema/establishment/repository/CareLineRoleRepository.java
@@ -18,7 +18,7 @@ public interface CareLineRoleRepository extends SGXAuditableEntityJPARepository<
 			"WHERE clr.careLineId = :careLineId " +
 			"AND clr.roleId = :roleId " +
 			"AND clr.deleteable.deleted IS FALSE")
-	Boolean alreadyExistsCareLineAssignedRole(@Param("careLineId") Integer careLineId, @Param("roleId") Integer roleId);
+	Boolean alreadyExistsCareLineAssignedRole(@Param("careLineId") Integer careLineId, @Param("roleId") Short roleId);
 
 	@Transactional(readOnly = true)
 	@Query(" SELECT COUNT(1) > 0 " +
